refactor(applicant): remove exception package
