+++ src/main/kotlin/apply/domain/applicant/ApplicantValidateException.kt
@@ -1,3 +1,3 @@
-package apply.domain.applicant.exception
+package apply.domain.applicant
 
 class ApplicantValidateException : RuntimeException("요청 정보가 기존 지원자 정보와 일치하지 않습니다")
